Make stream class public API
Summary:
Make stream class public API.

Stream should be a public class as it's API should be used by users e.g. `Stream.shape` ect. Users shouldn't directly create a `Steam` object but that's fine - an error is thrown if `__init__` is used. It is analogous to `Tensor` which is a public class that cannot be directly created by the user using `__init__`.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Tags: #popart.ir

Maniphest Tasks: T33421

Differential Revision: https://phabricator.sourcevertex.net/D62010