Remove several test cases from resize_tests.
Summary:
The resize tests are regularly failing CI with timeouts, and the amount of testing done for the resize ops is overkill. We should reduce the number of test cases run on resize, and add specific test cases if issues are encountered in the future.

We could probably remove more tests than this, but this is a good start.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Maniphest Tasks: T51472

Differential Revision: https://phabricator.sourcevertex.net/D62254