Add AiGraphcoreOpset1::slice to mimic Slice-1
Summary:
ONNX Slice v1 accepts axes, starts and ends as attributes rather
than inputs. This reduces the number of tensor constants and hence
ops in the ONNX model and can be used to reduce compile times. This
allows Slice-1 to be used via AiGraphcoreOpset1 and hence mixed
with later versions of ONNX.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, romand, markk, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, matthewha

Subscribers: matthewha, markk

Maniphest Tasks: T58902

Differential Revision: https://phabricator.sourcevertex.net/D63616