Allow ops in AOF to be absorbed into the gradient clipping cluster
Summary:
If an operation in the AOF is tied-constrainted to an operation in the
gradient clipping cluster, then the only way to guarantee that that operaiton can be scheduled correctly
is to include in the clipping cluster. This is relevant for merging collective operations, where certain
optimizer state collectives can already be performed during the gradient clipping stage.
For instance the cross-replica allreduce of LambR1Sq can be overlapped with the allreduce of the clipping-norm.
The allreduce, reduce-scatter, and allgather collectives required to construct the gradient were already
placed into the gradient clipping cluster as the gradient they produce is an input to the gradient clipping operations.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, fabiant, george

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: shengf

Tags: #mlperf_bert

Maniphest Tasks: T57645

Differential Revision: https://phabricator.sourcevertex.net/D63071