Allow for variable creation with VariableSettings in popXL
Summary:
This adds the conversion from the popXL `ReplicaGrouping` class to the popart `VariableSettings` class.

Eventually the variable settings are passed down to `addVarInit` which constructs an internal popart variable, which in turn creates a popXL variable.

Currently no functionality to return the ReplicaGrouping from a variable, this might be required in future.

Test Plan: Attempted some unit tests to verify the conversion works correctly. Integration tests to be added soon.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, johannesm, zigmasb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Subscribers: joshl, fabiant, oskarb

Maniphest Tasks: T54824, T56724

Differential Revision: https://phabricator.sourcevertex.net/D63339