Fix  'Unable to find inverse loss scale tensor' bug when applying the AutomaticLossScaling transform on some models
Summary: In a training graph, there are cases where variable tensors are not connected to an optimiser. Two cases we have identified are when the tensor is upstream from a DetachOp, and when the variable update type != VariableUpdateType::Gradient (e.g. for mean/variance inputs to a BatchNormOp). As such, these variable tensors do not have an associated inverse loss scaling tensor, so we should not try and search for one in such cases when applying the automatic loss scaling transform.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, arjunc, romand

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Subscribers: arjunc, evan

Maniphest Tasks: T55873

Differential Revision: https://phabricator.sourcevertex.net/D64122