Fix missing RTSGroup error
Summary:
When creating a RTS tensor (meta_shape is not empty) we need it's RTS Group to be initialised. This typically this happen during unwinding of the Tensor, however the lowering order of operations can cause this to have not happened yet..
The fix is to RTS group trace in `getInitTensorCreators` if needed.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: fabiant, shirazb

Maniphest Tasks: T60586

Differential Revision: https://phabricator.sourcevertex.net/D64986