.arcconfig: let arc search for linters
Summary:
arc has built-in search for linters, so let's not assume they exist in
the parent directory. arc will find them there in the CBT env anyway,
because looking in the parent directory is part of the search.

This change allows using (a preconfigured) arc without having to check
out the linters in the parent directory.

Ref T60316.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Maniphest Tasks: T60316

Differential Revision: https://phabricator.sourcevertex.net/D65281