ignore graph.isSchedulable to speedup applyInplacePattern
Summary:
the compilation time is very long, and ir's share is very big.
We found the ir's duration is about 1300 seconds in our model:
total of getSchedule is about 700 seconds
total of other components in ir is about 600 seconds
furthermore, we found the cost of "applyInplacePattern" is about 450 seconds.
and then, the biggest bottleneck is "graph.isSchedulable"
we read the reference code carefully, and we found it can be ignored if the "newTopoCons" is empty
and then, we checked the real situation,  we found the total number of op is 5920 when our code go into "applyInplacePattern",
and the number of  ops which its newTopoCons is not empty, is only about 850.
So, the total cost of ir is reduced to 950s after we adding this constraint

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, fabiant

Subscribers: matthewha, jamesn, hanzhao, xiaobingw

Maniphest Tasks: T59349

Differential Revision: https://phabricator.sourcevertex.net/D65355