Add context manager to popxl.Session that implictly manages attach, WeightsFromHost, WeightsToHost, and detach
Summary:
Depends on D65624.

We implement a context manager for `popxl.Session` and document it in the user guide.
We also remove the `weights_from_host` parameter of `popxl.Session.write_variable_data`. Always writing to device when attached and not doing so when unattached is a much more natural api.

We change the ctor to not detach (by passing `prepareDevice(loadEngine=False)`, delaying it until context enter.
On enter, the context manager will attach and `weights_from_host`.
On exit, the context manager will detach and `weights_to_host`.

It is important to detach on exit to reduce the chance of the user remaining attached throughout long running code that does not require a device. This makes it easier for many users sharing a machine.

Attaching/detaching can invalidate the memory on IPU, so it is important to `weights_from_host` on attach and `weights_to_host` on detach. This will ensure the API behaves coherently, e.g.:
```
for d in data:
  with session:
     session.run(d)
```
The attaching/detaching would invalidate the weights on device otherwise.

We force the user to be attached before calling `run`. This forces the user to use the context manager.

Alternatively, the user //can// still manually control attach/detach (and `weights_from_host`, `weights_to_host`) by directly accessing `session.device`, but this should not be required in all expected cases now.

The context manager can be nested. Nested enters will only set up the context (i.e. ensuring attach and calling `weights_from_host`) when going from detached to attached. Exits will only teardown the context (detach and `weights_to_host`) if the user was attached just before the corresponding enter. So:
```
assert not session.device.isAttached
with session:
  assert session.device.isAttached
  with session:
    pass
  # Still attached
  assert session.device.isAttached
# Now detached
assert not session.device.isAttached
```
In order to implement this, we need to store a stack of "was attached" states in `Session` and push/pop in enter/exit.

`get_tensor_data` and `write_variable_data` are both possible when detached (outside the context).
For `get_tensor_data`, we will return the current host buffers.
For `write_variable_data`, we will update only the host buffers, deferring the device transfer until the next context enter.

Test Plan: Add relevant integration tests for the context manager. Amend all existing tests to use the context manager.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, oskarb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, oskarb

Subscribers: matthewha, oskarb

Maniphest Tasks: T59969

Differential Revision: https://phabricator.sourcevertex.net/D65625