Add retrieval mode to replica grouping in PopXL
Summary:
Adds the retrieval modes:
- "one_per_group": Return only the first replica's variable per group.
- "all_replicas": Return all replica's variables in every group.

To the replica_grouping in PopXL. This is analogous to the VariableRetrivalMode in popart, except the unsupported `AllReduceReplicas` is not an option. See `willow/include/popart/variablesettings.hpp` for more info.

Also:
Reduced code in `Tensors::addVarInit` by relying on the shape on replica / shape on host functions to ensure the added tensor is the correct shape when using variable settings. This avoids the iflock code that is removed in this diff.
By similar logic `addVarInitWithLeadingGroupDim` is removed, as it's now superfluous given the shape on host / shape on replica functions.

Test Plan: Tests added to verify the shape of variables created and the conversion from PopXL to Popart. Numerical test added.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, zigmasb, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, zigmasb, shirazb

Subscribers: romand, shirazb, zigmasb, michaell, markk, johannesm

Maniphest Tasks: T59963

Differential Revision: https://phabricator.sourcevertex.net/D64629