Remove possibly redundant setTileMapping calls.
Summary:
There are two patterns in popart where a scalar variable is created, and the tile mapping is then set to 0.

1) A call to `getScalarTensor` is made, and then the default mapping is overriden with a call to `setTileMapping(x, 0);`.
2) In stash and restore, a scalar variable is created using `Graph::addVariable` and the tile mapping is set to 0.

In the first instance, it is unclear why the default tile mapping is overriden, and is probably unecessary.
The second instance is less clear. At first I thought that the author may not have known about `getScalarVariable`, but when I tried to use it, I got an error due to the shape being wrong (`{}` vs `{1}`). Directly creating the variable may have been because of this, but another solution is to call `reshape({1})` on the result of `getScalarVariable`, and then the call to `setTileMapping` may be ommited.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, matthewha, fabiant, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Maniphest Tasks: T62015

Differential Revision: https://phabricator.sourcevertex.net/D66717