Move GRU_Helper and LSTM_Helper classes to rnn_helper.py
Summary:
They should be kept in the same file as they are analogous to each other,
and we can use a single copy of the ONNX license. Additionally made sure that the
helper classes come from the ONNX 1.6.0 release, which is the one we use elsewhere.

Test Plan: Existing rnn, gru and lstm tests pass

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Maniphest Tasks: T55268

Differential Revision: https://phabricator.sourcevertex.net/D66783