Take compilation-affecting engine options into account when calculating hashes for the purpose of executable caching
Summary:
This diff ensures a hash of Poplar's engine options is taken into account when calculating the hash for the purposes of executable caching.

Note that before this change, engine options were derived by Poplar at engine creation time, taking into account the `OptionFlags` instance `IrLowering::engineOptions` (which populated in the `Devicex` constructor) as well as things like the environment variable `POPLAR_ENGINE_OPTIONS`. Unfortunately, this happens after PopART calculates a hash, and after the compilation process [that we're trying to cache by engine caching] has already happened, so it's too late. Also note that `poplar::EngineOptions` is something that Poplar uses under the hood, and is not passed in to an poplar::Engine at present, although this is set to change in T62481. So at present, PopART does not create an poplar::EngineOptions object.

* Move the logic that populates the `poplar::OptionFlags` instance that goes into an engine to somewhere that can be used earlier in the Session's lifetime (refactored this to `EngineOptionsCreator` class that can be re-used).
* Create an `poplar::EngineOptions` object from the `poplar::OptionFlags` and use the hash function defined by Poplar on the `poplar::EngineOptions` to include it in any hash calculation.
* Use the same logic when getting `poplar::OptionFlags` to set for engine creation / executable serialisation.
* Add tests to check that with changing engine options there is no cache hit for both PopART and PopXL.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T53239

Differential Revision: https://phabricator.sourcevertex.net/D67426