Add LABELS to CMake test helper functions
Summary:
This will allow passing `LABELS multicard` to have the `Hw` variant of the test
(if it exists) run on only the POD diff builder, not the C200 builder.

Actually, this was already implemented in `add_popart_cpp_unit_test` but not
documented, so we only add the documentation here.

In `add_popart_py_unit_test`, the API existed but was not implemented or
documented. We remedy both in this diff.

We change the `overlapio` test to use `LABELS multicard` to load balance across
the two hw builders, as it is a long test, and to be able to manually test this
diff.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Maniphest Tasks: T63409

Differential Revision: https://phabricator.sourcevertex.net/D68321