Fix loss scale update factore cache issue
Summary:
The problem was that the data of `lossScaleUpdateFactor` was not being serialised in the capnproto buffer. In ONNX models, tensors with initialisers get their data stored in the ONNX file. There's a piece of logic in `serializePopartExecutable` that would decide whether to store a tensor's data in the capnproto buffer if it doesn't have an initialiser OR it is an optimiser state tensor (snippet below).

The `lossScaleUpdateFactor` has an initialiser (added by `Ir::addAdditionalModelProtoTensors()`), so to add it to the capnproto buffer it also needs to be an optimiser state tensor.

```lang=cpp, name=willow/src/popx/executablexserializer.cpp
...
    for (auto &id : variableTensors) {
      Tensor *tensor = ir.getTensor(id);
      if (!tensor->hasProducer()) {
        auto tensorBuilder = tensors[i];

        // For Onnx-Ir models, we don't store the tensorData
        // for the variable tensors with initializers since
        // they will be loaded from the onnx file.
        // For Ir models, and others, the tensor data is always serialized
        bool serializeTensorData = true;
        if (ir.hasOnnxModel()) {
          bool isInitializer =
              popart::onnxutil::isInitializer(ir.getModel(), id);
          serializeTensorData =
              !isInitializer || tensor->isOptimizerStateTensor();
        }

        serializeTensor(tensor, tensorBuilder, serializeTensorData);
        ++i;
      }
    }
...
```

This diff fixes this by adding the `lossScaleUpdateFactor` to the optimiser state tensors.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, sergiop, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, sergiop, evan

Subscribers: evan, matthewha, romand, sergiop

Maniphest Tasks: T62581

Differential Revision: https://phabricator.sourcevertex.net/D69595