Clean up context even if entering the `Session` context manager fails
Summary:
Previously, if an exception occured in `Session.__enter__`, `Session.__exit__` would not be called. This could lead to a state where the context manager does not dettach from the device.

In this diff, we handle exceptions such that `Session.__exit__` gets called for clean up if entering the context manager fails.

Reviewers: shirazb, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha, oskarb

Maniphest Tasks: T63811

Differential Revision: https://phabricator.sourcevertex.net/D68978