Enable profiling of cached executables
Summary:
This diff enables profiling of cached executables.

If executable caching is enabled and a model is to be compiled, `profile.pop` and `debug.cbor` files will be copied from the `autoReport.directory` to the directory of the executable cache.

Before execution of a cached executable these files will be copied from the cached executable directory to the `autoReport.directory` where Poplar will append them.

As Poplar renames `profile.pop` to `profile_<random>.pop` when the files are written to, and since the `profile_<random>.pop` will be renamed back to `profile.pop` first when the `poplar::ProfileWriter` is destroyed, the storing and restoring needs to happen at specific places during `Devicex::prepare`.

Storing will happen just after the graph is compiled, and restoring will happen just before the creation of a `poplar::Engine` (the destruction of this object is what ultimately destroys `poplar::ProfileWriter`).

A separate class (`CachedExecutableProfiler`) has been made to deal with the copying of the files.

One current shortcoming is that specification of `autoReport.directory` needs to happen through `popart::SessionOptions` instead of through the `POPLAR_ENGINE_OPTIONS`  environmental variable. See T63870 for details

Test Plan: This diff contains unittests of the `CachedExecutableProfiler` together with an integration test testing profiling of cached executables.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Subscribers: simonl, markk, hectorm

Maniphest Tasks: T56614

Differential Revision: https://phabricator.sourcevertex.net/D68289