PopXL distributed data parallel support
Summary:
Support for distributed data parallel requires:
* Setting SessionOptions from `popdist`
* Acquiring device from `popdist`
This can be accessed by using a new argument on Ir construction `replication`. This can be set to an int for the `replication_factor` or `popdist` to setup the Ir for distributed data parallel. A new property of Ir is added `instance_replication_factor`. This is then used in the `Session` for the purposes of validating/constructing the input/output tensors of `run`.

To be able to use RTS with multi-instance programs the sharding must be limited to within an instance. Changes to help support this:
* New api method - `replica_sharded_buffer` - helps create the correct sized buffer.
* New argument on `replica_sharded_variable` - `sharding_factor`
* New virtual method `CollectiveOp::getCommSize` - this allows ops to specify how many replicas the CBR should be instantiated with. This allows popxl strided collectives to specify a scheme that is not representable by CommGroup
* Validate that all RTS ops in a RTSGroup have the same `CommSize`
* Modify `weightsToHost` and `weightsFromHost` to use the CBR `getReplicationFactor` for calculating rearrangement addresses.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand

Subscribers: romand, markk, jamesn

Maniphest Tasks: T64427

Differential Revision: https://phabricator.sourcevertex.net/D68759