Remove the CIFAR /torch tests from popart.
Summary:
As per the task, these tests are unnecessary and are replaced by equivalent tests in pytorch. They are also quite flakey, especially considering the MNIST dataset download.

As per comment on T64464 this was not the cause of the CI timeout so I am going to re-land and see if the issue persists.

Test Plan: CI

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Subscribers: markk, jamesn

Maniphest Tasks: T64465

Differential Revision: https://phabricator.sourcevertex.net/D69861