Fix for replica grouping and writing variables
Summary:
Currently using `write_variables_data` with replica grouping will produce an error. This diff fixes this error and allows for copying to sharded variables via `write_variables_data`.

Unfortunately I could not find a solution to copy to variables with "all_replicas" retrieval mode, whilst this is annoying, since it is only used for testing I propose we leave this as-is and add a task for it.

Test Plan: Various tests added.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #popart, joshl, romand

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #popart, joshl, romand

Subscribers: romand, joshl, fangliangs

Maniphest Tasks: T61041

Differential Revision: https://phabricator.sourcevertex.net/D68789