Introduce API into the session class to allow obtaining all tensor IDs
Summary:
Ref T56357

Poptorch allows obtaining tensor names from the model.
It is needed for anchoring API. It was getting info about
that from serialized IR. It was impossible when if a user
created a session from the cached executable. To fix that
issue popart can give a user possibility to get all tensor
ids from the model. Besides documentation for python bindings was regenerated.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: yanislavd, tomb, shirazb, markk

Maniphest Tasks: T56357

Differential Revision: https://phabricator.sourcevertex.net/D64173