Skip verifyPadBase when tensor shape not known
Summary:
The tensor shape does not need to be known by the builder for the
for the model to function, as PopART's own shape inference can
infer the shape. However, verifyPadBase will fail. With this patch,
the test will be skipped, as happens in other verificayion tests,
if the tensor shape is not known.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T66303

Differential Revision: https://phabricator.sourcevertex.net/D71606