Ensuring order of "replacedDebugContextIds" matches "outlinedDebugContextIds" in Call debug info.
Summary: Before, it was possible the order of each array would be different. Now we produce the arrays with the same logic as `opRemaps` to ensure the order is correct.

Test Plan: The one unit test that checks these fields happen to be already ordered correctly. I am not sure how to create a test case where they would've been incorrectly ordered.

Reviewers: markk, leiw, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell

Reviewed By: leiw, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell

Subscribers: michaell

Maniphest Tasks: T66465

Differential Revision: https://phabricator.sourcevertex.net/D71780