Move popart to use the gccs python unit test file
Summary:
See D71279 and D70959. This task is to make the final switch from the AddPythonUnitTest.cmake file in popart to the one in GCCS.

Only 1 additional change was made, which was to add the python3 path as an optional input to the function, so I have modified that here.

Test Plan: CI

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Maniphest Tasks: T67352

Differential Revision: https://phabricator.sourcevertex.net/D72921