Adding POPART_LAYER_REGEX_FOR_OUTLINER environment variable to manipulate outliner behaviour (for use with PopLiner).
Summary:
This is my suggestion on how to overcome the overly aggressive outlining that occurs when compiling a large model on a single IPU. This diff allows the user to instruct the PopART outliner to treat operations from different layers as if they are on different virtual graphs (and therefore, IPUs) - making them far less likely to be combined during outlining. This is necessary for PopLiner as it gives a more typical representation of the final graph when compiled for multiple IPUs.

I am open to suggestions on better ways to tackle this problem.

Test Plan: I have tested this with PopTorch Bert Base/Large Squad and have checked the debug logging output as (to see the correct layers are found) well as the profile in Graph Analyser (to see that all input variables for each CallOp are from the same layer).

Reviewers: simonl, markk, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: markk, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: matthewha

Maniphest Tasks: T63773

Differential Revision: https://phabricator.sourcevertex.net/D72576