When constructing a popxl.Ir, with replication='popdist', instruct PopDist to acquire the device with OnDemand conection type
Summary:
Depends on D73598.

In PopXL, we would like to have the behaviour that constructing a `Session`
and asking PopXL to find a device for you (so whenever you do not pass in a
device you have already acquired, but a string like `'ipu_hw'`), the ctor does
not attach to the device; only entering the context does.

When you pass an `Ir` to a `Session` that was constructed with
`replication='popdist'`, the `Session` ctor will ask `popdist.popart.getDevice`
to get the device for you. This function acquires the device using
`popart.DeviceConnectionType.Always`, meaning the device is immediately
acquired.

This diff uses the new PopDist API in D73598 to pass
`popart.DeviceConnectionType.OnDemand` to `popdist.popart.getDevice`, so the
desired API of not attaching in the ctor is maintained.

Test Plan:
We extend the test of the `Session`'s attach-detach behaviour to
cover the `replication='popdist'` case.

This test fails on CI for an unknown reason. Have left the test
disabled with a follow-up task for fixing.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, yanislavd

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, yanislavd

Subscribers: yanislavd

Maniphest Tasks: T68244

Differential Revision: https://phabricator.sourcevertex.net/D73632