Adding none reduction to popart scatterreduce op
Summary:
This adds a new `None` reduction method to target from the `ScatterReduceOp`.  This allows us to reuse the logic for handling both the cases where we are doing:

* elementwise scatter where we are moving scalar values from one tensor to another (now called ScatterElements in onnx)
* vectorised scatter where we are assigning whole vectors from one tensor to another (a common use case, now defined as ScatterND in onnx)

The attached profile shows a comparison between the onnx scatter implementation and an equivalent `none` scatter reduce which shows:

* 100x reduction in cycle count
* significantly better memory profile

Additionally the operator interface now supports providing an optional third input that is used as the initial value for the scatter operation.  This follows the signature of the upstream [[ https://pytorch-scatter.readthedocs.io/en/latest/functions/scatter.html | torch scatter ]] library that we support through poptorch.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, yanislavd

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, yanislavd

Subscribers: yanislavd

Maniphest Tasks: T49876, T61957

Differential Revision: https://phabricator.sourcevertex.net/D74229