Prepare for change from gcl_ct.so to gcl.so
Summary:
The aim is that the GCL library will be named gcl.so in SDK 3.0, so
support both for now.

Note: The Findgcl.cmake file removal is in progress, but will most
likely not be complete in the 3.0 timeframe. GCL supports its removal,
but then the dependency issues hidden by the file rears their ugly
heads; this is a hydra-like problem where fixing one issue triggers
two new issues.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, aysanm, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, aysanm, matthewha

Maniphest Tasks: T62930

Differential Revision: https://phabricator.sourcevertex.net/D74433