Automatically manage the popdist backend for the user
Summary:
Add file `popdist_backend_manager` with function `init` that calls into
`popdist.init` only the first time the function is called during the program,
else does nothing.

Additionally, it has an option that, if the `popdist` backend was not already
initialised by the user, will print a deprecation warning, as in the future we
will no longer do this for the user.

If `enableDistributedReplicatedGraphs`, we call this function in the `popart`
Python `Session` classes.
On the PopXL path, we do not print this warning. On the ONNX PopART path,
we do.

Note, even if the user did not use popdist to configure the replication factor
and devices used (either by manually calling popdist, or in PopXL by passing
`replication='popdist'`), PopART internally will use the popdist library to perform
inter-instance communication, and thus we still need to setup the backend.

Test Plan:
ONNX path manually tested, as multi-instance tests are disabled on CI.
PopXL path also manually tested, using tests from D73887.

Reviewers: oskarb, nejci, yanislavd, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, christiana

Reviewed By: yanislavd, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: nejci

Maniphest Tasks: T68253

Differential Revision: https://phabricator.sourcevertex.net/D73788