Support VariableSettings and RTS for remote variables on multi-instance programs, for PopXL only
Summary:
Depends on D73788.

When there are multiple host instances, each instance only has access to its
local subset of replicas. Therefore, depending on the replica grouping, sharding
domain size, and variable retrieval mode, the instances may need to communicate
in weightsToHost to recover the full global weight data from the other
instances.

This diff implements this //only for remote variables, not on-chip//.

`Devicex::remoteBufferWeightsFromHost` and `Devicex::remoteBufferWeightsToHost`
are responsible for copying between the local replicas' remote buffers and the
corresponding slices of the weight's `TensorData`/`d2hWeightBuffer`. Previously,
the logic in these functions did not at all handle multiple instances. This diff
updates the logic to consider when there are multiple instances. Secondly, this
diff performs host-side collectives using the popdist library in
`remoteBufferWeightsToHost` to recover the full weight data on each instance.
Inside of popdist, these are MPI collectives.

**The user API is that they must provide the full weight data on each instance,
and they will be returned the full weight data on each instance.**

Test Plan:
Comprehensive integration tests through the PopXL API. Note these are checked
manually and not on CI due to limitations of CI.

Check that existing tests through PopART API are not broken. This diff does not
attempt to test that the //new// features work for PopART, only PopXL, though
there is no reason they shouldn't. Again this is manually tested as the tests do
not run on CI.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oskarb, johannesm, fabiant, yanislavd

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, yanislavd

Maniphest Tasks: T68254

Differential Revision: https://phabricator.sourcevertex.net/D73887