Use poplar::ReplicatedStreamMode::REPLICATE when there is >1 groups, calculated from GLOBAL replicas not local
Summary:
Fixes T69645.

In Poplar, `::REPLICATE` should be used whenever there is more than 1 //global//
replicas and we are not broadcasting the same data to every replica.

Therefore, in PopART, we should calculate the number of groups in the replica
grouping based on the global replication factor, not local, then check that
number of groups is greater than 1. Even if there is no local replication in
PopART, there is still (global) replication in Poplar, and we are sending
different data to the replicas (in different groups), so `::REPLICATE` should
be used.

Also, remove a dead if-else branch that repeated the same condition as the
previous branch so was never hit.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Maniphest Tasks: T69645

Differential Revision: https://phabricator.sourcevertex.net/D75422