Add `conv_pow2scaled` op to popxl
Summary:
The existing Conv and ConvX classes are extended to handle Float8 inputs - that is, no new PopART ops are added.
To perform a float8 convolution, the user must pass in float8 tensors for the inputs and weight, and an int32 log2scale tensor.
This only adds support for the Conv op - float8 conv transpose and multiconv are not yet supported.
Trying to get the grad ops of a float8 convolution will result in error.

Depends on D77126

Test Plan: Tests for convop have been extended with some new cases and new popxl integration test added. The latter tests against Pytorch's cpu conv implementation to check for numerical correctness.

Reviewers: markk, matthewha, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: matthewha, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T69673

Differential Revision: https://phabricator.sourcevertex.net/D77274